+++ app/src/main/java/alesk/com/masterhelper/domain/repository/ProjectsRepository.kt
@@ -4,8 +4,8 @@ import alesk.com.masterhelper.data.entities.Project
 
 interface ProjectsRepository {
     fun getAllProjects(): List<Project>
-    fun getProject(id: Int): Project?
+    fun getProject(id: Long): Project?
     fun createProject(project: Project)
     fun updateProject(project: Project)
-    fun deleteProject(id: Int)
+    fun deleteProject(id: Long)
 }
