+++ app/src/main/java/alesk/com/masterhelper/presentation/main/MainRouter.kt
@@ -3,6 +3,6 @@ package alesk.com.masterhelper.presentation.main
 interface MainRouter {
     fun showProjectsList()
     fun showCreateNewProject()
-    fun showProjectInfo(id: Int)
+    fun showProjectInfo(id: Long)
     fun showMasterInfo()
 }
